# agri_uav_disease_system/backend/docs/integration_package/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "三下乡无人机水稻病虫害识别系统",
    "version": "mvp-1"
  },
  "paths": {
...
    "/api/experimental/uav-blb-segmentation/dry-run": {
      "post": {
        "tags": [
          "experimental-uav-blb-segmentation"
        ],
        "summary": "Uav Blb Segmentation Dry Run",
        "operationId": "uav_blb_segmentation_dry_run_api_experimental_uav_blb_segmentation_dry_run_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uav_blb_segmentation_dry_run_api_experimental_uav_blb_segmentation_dry_run_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UavBlbSegmentationDryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experimental/uav-blb-segmentation/field-trial": {
      "post": {
        "tags": [
          "experimental-uav-blb-segmentation"
        ],
        "summary": "Uav Blb Segmentation Field Trial",
        "operationId": "uav_blb_segmentation_field_trial_api_experimental_uav_blb_segmentation_field_trial_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uav_blb_segmentation_field_trial_api_experimental_uav_blb_segmentation_field_trial_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UavBlbSegmentationFieldTrialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experimental/uav-blb-segmentation/field-trial/records": {
      "get": {
        "tags": [
          "experimental-uav-blb-segmentation"
        ],
        "summary": "Uav Blb Segmentation Field Trial Records",
        "operationId": "uav_blb_segmentation_field_trial_records_api_experimental_uav_blb_segmentation_field_trial_records_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UavBlbSegmentationFieldTrialRecordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fields": {
      "post": {
        "tags": [
          "fields"
        ],
        "summary": "Create Field",
        "operationId": "create_field_api_fields_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "List Fields",
        "operationId": "list_fields_api_fields_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fields/{field_id}": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "Get Field",
        "operationId": "get_field_api_fields__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "fields"
        ],
        "summary": "Update Field",
        "operationId": "update_field_api_fields__field_id__put",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fields"
        ],
        "summary": "Archive Field",
        "operationId": "archive_field_api_fields__field_id__delete",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk/fusion/field/{field_id}": {
      "get": {
        "tags": [
          "risk-fusion"
        ],
        "summary": "Field Risk Fusion History",
        "operationId": "field_risk_fusion_history_api_risk_fusion_field__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskFusionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inspection-reports/generate": {
      "post": {
        "tags": [
          "inspection-reports"
        ],
        "summary": "Generate Report",
        "operationId": "generate_report_api_inspection_reports_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionReportGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inspection-reports": {
      "get": {
        "tags": [
          "inspection-reports"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_inspection_reports_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inspection-reports/{report_id}": {
      "get": {
        "tags": [
          "inspection-reports"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_inspection_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_uav_blb_segmentation_dry_run_api_experimental_uav_blb_segmentation_dry_run_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "return_probability_map": {
            "type": "boolean",
            "title": "Return Probability Map",
            "default": true
          },
          "return_overlay": {
            "type": "boolean",
            "title": "Return Overlay",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file",
          "mode"
        ],
        "title": "Body_uav_blb_segmentation_dry_run_api_experimental_uav_blb_segmentation_dry_run_post"
      },
      "Body_uav_blb_segmentation_field_trial_api_experimental_uav_blb_segmentation_field_trial_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "plot_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Name"
          },
          "operator_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Note"
          },
          "human_review_status": {
            "type": "string",
            "title": "Human Review Status",
            "default": "pending"
          },
          "human_review_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Review Label"
          },
          "issue_tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Tags"
          },
          "return_probability_map": {
            "type": "boolean",
            "title": "Return Probability Map",
            "default": true
          },
          "return_overlay": {
            "type": "boolean",
            "title": "Return Overlay",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file",
          "mode"
        ],
        "title": "Body_uav_blb_segmentation_field_trial_api_experimental_uav_blb_segmentation_field_trial_post"
      },
...
      "FieldCreate": {
        "properties": {
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "location_city": {
            "type": "string",
            "title": "Location City",
            "default": "宿迁市"
          },
          "location_district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location District"
          },
          "location_town": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Town"
          },
          "location_village": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Village"
          },
          "center_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Center Lat"
          },
          "center_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Center Lng"
          },
          "area_estimate_mu": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Estimate Mu"
          },
          "crop_type": {
            "type": "string",
            "title": "Crop Type",
            "default": "rice"
          },
          "current_growth_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Growth Stage"
          },
          "field_status": {
            "type": "string",
            "title": "Field Status",
            "default": "active"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "field_id": {
            "type": "string",
            "title": "Field Id"
          }
        },
        "type": "object",
        "required": [
          "field_name",
          "field_id"
        ],
        "title": "FieldCreate"
      },
...
      "FieldListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FieldInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "FieldListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InspectionReportGenerateRequest": {
        "properties": {
          "field_id": {
            "type": "string",
            "title": "Field Id"
          },
          "uav_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uav Task Id"
          },
          "include_rag": {
            "type": "boolean",
            "title": "Include Rag",
            "default": true
          },
          "include_risk": {
            "type": "boolean",
            "title": "Include Risk",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "field_id"
        ],
        "title": "InspectionReportGenerateRequest"
      },
...
      "RiskFusionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RiskFusionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RiskFusionListResponse"
      },
...
      "UavBlbSegmentationDryRunResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "dry_run_only"
          },
          "production_ready": {
            "type": "boolean",
            "title": "Production Ready",
            "default": false
          },
          "backend_integration_allowed": {
            "type": "string",
            "title": "Backend Integration Allowed",
            "default": "dry_run_only"
          },
          "model_stage": {
            "type": "string",
            "title": "Model Stage",
            "default": "formal_candidate"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "input_config": {
            "type": "string",
            "title": "Input Config",
            "default": "D2_5BAND_NDVI"
          },
          "weight_sha256": {
            "type": "string",
            "title": "Weight Sha256"
          },
          "patch_size": {
            "type": "integer",
            "title": "Patch Size",
            "default": 256
          },
          "stride": {
            "type": "integer",
            "title": "Stride",
            "default": 128
          },
          "threshold": {
            "type": "number",
            "title": "Threshold",
            "default": 0.45
          },
          "min_area": {
            "type": "integer",
            "title": "Min Area",
            "default": 128
          },
          "disease_area_ratio": {
            "type": "number",
            "title": "Disease Area Ratio"
          },
          "mask_url": {
            "type": "string",
            "title": "Mask Url"
          },
          "overlay_url": {
            "type": "string",
            "title": "Overlay Url"
          },
          "probability_map_url": {
            "type": "string",
            "title": "Probability Map Url"
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "default": "experimental_dry_run_only_not_for_production"
          },
          "original_preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Preview Url"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "weight_sha256",
          "disease_area_ratio",
          "mask_url",
          "overlay_url",
          "probability_map_url"
        ],
        "title": "UavBlbSegmentationDryRunResponse"
      },
...
      "UavBlbSegmentationFieldTrialRecordsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "field_trial_only"
          },
          "production_ready": {
            "type": "boolean",
            "title": "Production Ready",
            "default": false
          },
          "backend_integration_allowed": {
            "type": "string",
            "title": "Backend Integration Allowed",
            "default": "field_trial_only"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/UavBlbSegmentationFieldTrialRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "records",
          "total"
        ],
        "title": "UavBlbSegmentationFieldTrialRecordsResponse"
      },
      "UavBlbSegmentationFieldTrialResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "field_trial_only"
          },
          "production_ready": {
            "type": "boolean",
            "title": "Production Ready",
            "default": false
          },
          "backend_integration_allowed": {
            "type": "string",
            "title": "Backend Integration Allowed",
            "default": "field_trial_only"
          },
          "model_stage": {
            "type": "string",
            "title": "Model Stage",
            "default": "formal_candidate"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "input_config": {
            "type": "string",
            "title": "Input Config",
            "default": "D2_5BAND_NDVI"
          },
          "weight_sha256": {
            "type": "string",
            "title": "Weight Sha256"
          },
          "patch_size": {
            "type": "integer",
            "title": "Patch Size",
            "default": 256
          },
          "stride": {
            "type": "integer",
            "title": "Stride",
            "default": 128
          },
          "threshold": {
            "type": "number",
            "title": "Threshold",
            "default": 0.45
          },
          "min_area": {
            "type": "integer",
            "title": "Min Area",
            "default": 128
          },
          "disease_area_ratio": {
            "type": "number",
            "title": "Disease Area Ratio"
          },
          "mask_url": {
            "type": "string",
            "title": "Mask Url"
          },
          "overlay_url": {
            "type": "string",
            "title": "Overlay Url"
          },
          "probability_map_url": {
            "type": "string",
            "title": "Probability Map Url"
          },
          "original_preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Preview Url"
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "default": "field_trial_not_for_production"
          },
          "trial_record": {
            "$ref": "#/components/schemas/UavBlbSegmentationFieldTrialRecord"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "weight_sha256",
          "disease_area_ratio",
          "mask_url",
          "overlay_url",
          "probability_map_url",
          "trial_record"
        ],
        "title": "UavBlbSegmentationFieldTrialResponse"
      },
...
    }
  }
}